COMP: Replacing "unsigned long" with "size_t" as the basic IdentifierType.       In this way we should get a consistent behavior for large number of       samples, when using 64bits platforms.

+++ Code/Review/Statistics/itkMeasurementVectorTraits.h
@@ -94,8 +94,7 @@ public:
       precision). Using double type will also ultimately fall into the same problem.
       Hence in the new statistics framework, InstanceIdentifier/FrequencyTypes are 
       set to the the largest possible integer on the machine */  
-  //typedef typename std::vector<int>::size_type   InstanceIdentifier;
-  typedef unsigned long  InstanceIdentifier;
+  typedef size_t      InstanceIdentifier;
 
   /** Type defined for representing the frequency of measurement vectors */
   typedef InstanceIdentifier                                      AbsoluteFrequencyType;
